* Moved test Client implementation to utils inside user_sessions so it is available to users installing the egg only * Changed comments format to suit the project's requirements
